backend/requirements.txt
flask
flask-cors
opencv-python
mediapipe
deepface 
numpy
pandas 
matplotlib 
librosa 
pyaudio 
torch
tf-keras
SpeechRecognition
fastapi
uvicorn
groq
python-dotenv
requests
gunicorn
